feat: drop support for vim, neovim <= 0.4
This commit drops support for vim and neovim 0.4.x. neovim 0.5.x. neovim
0.5 (nightly) is generally stable enough for daily use and introduces
several new APIs (e.g. the LSP and Treesitter APIs). These APIs, in
conjunction with a few new plugins that use them, represent such an
improvement in everyday use that I'm willing to deal with the periodic
instability that comes with running nightly software.

I'm still making an effort to degrade gracefully on neovim 0.4 and
vanilla vim, but in reality I haven't used either for 6+ months, so it's
best effort only.