feat: use treesitter for JS, JSON syntax highlighting
treesitter is a framework for incrementally and performantly parsing
files to an AST, which is then provided to a consumer (in this case, the
vim text editor). It results in much better syntax highlighting than
both vim's built-in highlighting and the syntax file provided by
vim-polyglot; very anecdotally, it also seems to be more performant.

treesitter is still unstable software, so this commit only enables it
for the JavaScript and JSON filetypes. Assuming it works well, I'll
expand that filetype list over time.